+++ src/pages/landing/ebooks/forex.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import Introduction from 'components/layout/ebooks/introduction'
-import HeaderSection from 'components/layout/ebooks/header'
-import Topics from 'components/layout/ebooks/topics'
+import Introduction from './components/introduction'
+import HeaderSection from './components/header'
+import Topics from './components/topics'
 import Layout from 'components/layout/layout'
 import { SEO } from 'components/containers'
 import { localize, WithIntl } from 'components/localization'
