+++ src/pages/landing/ebooks/stocks.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import HeaderSection from 'components/layout/ebooks/header'
-import Introduction from 'components/layout/ebooks/introduction'
-import Topics from 'components/layout/ebooks/topics'
+import HeaderSection from './components/header'
+import Introduction from './components/introduction'
+import Topics from './components/topics'
 import Layout from 'components/layout/layout'
 import { SEO } from 'components/containers'
 import { localize, WithIntl } from 'components/localization'
